requirements.txt
disnake>=2.9.0
SQLAlchemy>=2.0.0
sqlmodel>=0.0.16
asyncpg>=0.29.0
alembic>=1.13.0
python-dotenv>=1.0.0
redis>=5.0.0
nanoid>=2.0.0
aiofiles>=23.2.1
Pillow>=10.3.0